+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/TabbedPanelLayout.java
@@ -35,6 +35,7 @@ import java.awt.Window;
 import java.awt.event.ActionEvent;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.util.Collection;
 import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.function.Function;
@@ -116,7 +117,7 @@ public final class TabbedPanelLayout implements PanelLayout {
 
   private EntityPanel entityPanel;
   private JTabbedPane detailPanelTabbedPane;
-  private JSplitPane tableDetailSplitPane;
+  private JSplitPane detailPanelSplitPane;
   private Window detailPanelWindow;
   private PanelState detailPanelState = EMBEDDED;
   private final boolean includeDetailTabPane;
@@ -135,18 +136,14 @@ public final class TabbedPanelLayout implements PanelLayout {
 
   @Override
   public void updateUI() {
-    Utilities.updateUI(detailPanelTabbedPane, tableDetailSplitPane);
+    Utilities.updateUI(detailPanelTabbedPane, detailPanelSplitPane);
   }
 
   @Override
   public void layout(EntityPanel entityPanel) {
     this.entityPanel = requireNonNull(entityPanel);
-    tableDetailSplitPane = createTableDetailSplitPane();
-    detailPanelTabbedPane = createDetailTabbedPane();
     entityPanel.setLayout(borderLayout());
-    entityPanel.add(tableDetailSplitPane == null ?
-            entityPanel.editControlTablePanel() :
-            tableDetailSplitPane, BorderLayout.CENTER);
+    entityPanel.add(createCenterComponent(), BorderLayout.CENTER);
     setupResizing();
     setupControls();
     initializeDetailPanelState();
@@ -244,51 +241,49 @@ public final class TabbedPanelLayout implements PanelLayout {
   }
 
   private void initializeDetailPanelState() {
-    if (detailPanelTabbedPane != null) {
-      Value<PanelState> detailPanelStateValue = detailController.panelState(selectedDetailPanel());
+    selectedDetailPanel().ifPresent(selectedDetailPanel -> {
+      Value<PanelState> detailPanelStateValue = detailController.panelState(selectedDetailPanel);
       if (detailPanelStateValue.isNotEqualTo(detailPanelState)) {
         detailPanelStateValue.set(detailPanelState);
       }
       else {
         detailController.updateDetailState();
       }
-    }
+    });
   }
 
-  private EntityPanel selectedDetailPanel() {
-    if (detailPanelTabbedPane == null) {
-      throw new IllegalStateException("No detail panels available");
-    }
-
-    return (EntityPanel) detailPanelTabbedPane.getSelectedComponent();
+  private Optional<EntityPanel> selectedDetailPanel() {
+    return Optional.ofNullable(detailPanelTabbedPane == null ? null : (EntityPanel) detailPanelTabbedPane.getSelectedComponent());
   }
 
-  private JSplitPane createTableDetailSplitPane() {
-    if (!includeDetailTabPane || entityPanel.detailPanels().isEmpty()) {
-      return null;
+  private JComponent createCenterComponent() {
+    if (includeDetailTabPane && !entityPanel.detailPanels().isEmpty()) {
+      detailPanelSplitPane = createTableDetailSplitPane(entityPanel.editControlTablePanel());
+      detailPanelTabbedPane = createDetailTabbedPane(entityPanel.detailPanels());
+
+      return detailPanelSplitPane;
     }
 
+    return entityPanel.editControlTablePanel();
+  }
+
+  private JSplitPane createTableDetailSplitPane(JPanel editControlTablePanel) {
     return splitPane()
             .orientation(JSplitPane.HORIZONTAL_SPLIT)
             .continuousLayout(true)
             .oneTouchExpandable(true)
             .dividerSize(GAP.get() * 2)
             .resizeWeight(splitPaneResizeWeight)
-            .leftComponent(entityPanel.editControlTablePanel())
-            .rightComponent(detailPanelTabbedPane)
+            .leftComponent(editControlTablePanel)
             .build();
   }
 
-  private JTabbedPane createDetailTabbedPane() {
-    if (!includeDetailTabPane || entityPanel.detailPanels().isEmpty()) {
-      return null;
-    }
-
+  private JTabbedPane createDetailTabbedPane(Collection<EntityPanel> detailPanels) {
     TabbedPaneBuilder builder = tabbedPane()
             .focusable(false)
-            .changeListener(e -> selectedDetailPanel().activate())
+            .changeListener(e -> selectedDetailPanel().ifPresent(EntityPanel::activate))
             .onBuild(tabbedPane -> tabbedPane.setFocusCycleRoot(true));
-    entityPanel.detailPanels().forEach(detailPanel -> builder.tabBuilder(detailPanel.caption().get(), detailPanel)
+    detailPanels.forEach(detailPanel -> builder.tabBuilder(detailPanel.caption().get(), detailPanel)
             .toolTipText(detailPanel.getDescription())
             .add());
     if (includeDetailPanelControls) {
@@ -325,7 +320,7 @@ public final class TabbedPanelLayout implements PanelLayout {
 
     private static void resizePanel(EntityPanel panel, EntityPanel.Direction direction) {
       TabbedPanelLayout detailPanelLayout = panel.panelLayout();
-      JSplitPane splitPane = detailPanelLayout.tableDetailSplitPane;
+      JSplitPane splitPane = detailPanelLayout.detailPanelSplitPane;
       switch (requireNonNull(direction)) {
         case RIGHT:
           if (splitPane != null) {
@@ -348,13 +343,14 @@ public final class TabbedPanelLayout implements PanelLayout {
 
     @Override
     public void mouseReleased(MouseEvent e) {
-      EntityPanel detailPanel = selectedDetailPanel();
-      if (e.getClickCount() == 2 && e.getButton() == MouseEvent.BUTTON1) {
-        detailController.panelState(detailPanel).map(panelState -> panelState == WINDOW ? EMBEDDED : WINDOW);
-      }
-      else if (e.getButton() == MouseEvent.BUTTON2) {
-        detailController.panelState(detailPanel).map(panelState -> panelState == EMBEDDED ? HIDDEN : EMBEDDED);
-      }
+      selectedDetailPanel().ifPresent(selectedDetailPanel -> {
+        if (e.getClickCount() == 2 && e.getButton() == MouseEvent.BUTTON1) {
+          detailController.panelState(selectedDetailPanel).map(panelState -> panelState == WINDOW ? EMBEDDED : WINDOW);
+        }
+        else if (e.getButton() == MouseEvent.BUTTON2) {
+          detailController.panelState(selectedDetailPanel).map(panelState -> panelState == EMBEDDED ? HIDDEN : EMBEDDED);
+        }
+      });
     }
   }
 
@@ -372,6 +368,7 @@ public final class TabbedPanelLayout implements PanelLayout {
 
     @Override
     public void select(EntityPanel detailPanel) {
+      requireNonNull(detailPanel);
       if (detailPanelTabbedPane != null) {
         detailPanelTabbedPane.setFocusable(true);
         detailPanelTabbedPane.setSelectedComponent(detailPanel);
@@ -388,50 +385,35 @@ public final class TabbedPanelLayout implements PanelLayout {
     }
 
     private void updateDetailState() {
-      if (detailPanelTabbedPane == null) {
-        return;
-      }
-
-      PanelState previousPanelState = previousPanelState();
-      if (panelState.isNotEqualTo(HIDDEN)) {
-        selectedDetailPanel().initialize();
-      }
-
-      if (previousPanelState == WINDOW) {//if we are leaving the WINDOW state, hide all child detail windows
-        for (EntityPanel detailPanel : entityPanel.detailPanels()) {
-          detailPanel.panelLayout().detailController()
-                  .filter(TabbedDetailController.class::isInstance)
-                  .map(TabbedDetailController.class::cast)
-                  .ifPresent(controller -> {
-                    if (controller.panelState.isEqualTo(WINDOW)) {
-                      controller.panelState.set(HIDDEN);
-                    }
-                  });
+      selectedDetailPanel().ifPresent(selectedDetailPanel -> {
+        if (panelState.isNotEqualTo(HIDDEN)) {
+          selectedDetailPanel.initialize();
         }
+        SwingEntityModel selectedDetailModel = selectedDetailPanel.model();
+        if (entityPanel.model().containsDetailModel(selectedDetailModel)) {
+          entityPanel.model().detailModelLink(selectedDetailModel).active().set(panelState.isNotEqualTo(HIDDEN));
+        }
+      });
+      if (previousPanelState() == WINDOW) {
         disposeDetailWindow();
       }
-
-      SwingEntityModel detailModel = selectedDetailPanel().model();
-      if (entityPanel.model().containsDetailModel(detailModel)) {
-        entityPanel.model().detailModelLink(detailModel)
-                .active().set(panelState.isNotEqualTo(HIDDEN));
-      }
       if (panelState.isEqualTo(EMBEDDED)) {
-        if (tableDetailSplitPane.getRightComponent() != detailPanelTabbedPane) {
-          tableDetailSplitPane.setRightComponent(detailPanelTabbedPane);
-        }
+        detailPanelSplitPane.setRightComponent(detailPanelTabbedPane);
       }
       else if (panelState.isEqualTo(HIDDEN)) {
-        tableDetailSplitPane.setRightComponent(null);
+        detailPanelSplitPane.setRightComponent(null);
       }
       else {
-        showDetailWindow();
+        displayDetailWindow();
       }
 
       entityPanel.revalidate();
     }
 
     private PanelState previousPanelState() {
+      if (detailPanelTabbedPane == null) {
+        throw new IllegalStateException("No tabbed detail pane available");
+      }
       if (detailPanelWindow != null) {
         return WINDOW;
       }
@@ -491,7 +473,7 @@ public final class TabbedPanelLayout implements PanelLayout {
       return controls.build();
     }
 
-    private void showDetailWindow() {
+    private void displayDetailWindow() {
       Window parent = parentWindow(entityPanel);
       if (parent != null) {
         Dimension parentSize = parent.getSize();